better naming? re-organize the api middleware to avoid apps logic
